+++ src/deposit.h
@@ -1,6 +1,6 @@
 #ifndef DEPOSIT_H
 #define DEPOSIT_H
 
-int analiz (int d, int s);
+int analiz (float d, int s);
 
 #endif
